(feat) - reimplement the query resolvers
