+++ src/operations/query.ts
@@ -1,5 +1,3 @@
-import warning from 'warning';
-
 import {
   getFragments,
   getMainOperation,
@@ -14,7 +12,6 @@ import {
   Fragments,
   Variables,
   Data,
-  DataField,
   Link,
   SelectionSet,
   Completeness,
@@ -107,35 +104,7 @@ const readSelection = (
 
     if (isQuery) addDependency(fieldKey);
 
-    const resolvers = store.resolvers[typename];
-    if (resolvers !== undefined && resolvers.hasOwnProperty(fieldName)) {
-      // We have a resolver for this field.
-      const resolverValue = resolvers[fieldName](
-        data,
-        fieldArgs || {},
-        store,
-        ctx
-      );
-
-      if (node.selectionSet === undefined) {
-        // If it doesn't have a selection set we have resolved a property.
-        // We assume that a resolver for scalar values implies that this
-        // field is always present, so completeness won't be set to EMPTY here
-        data[fieldAlias] = resolverValue !== undefined ? resolverValue : null;
-      } else {
-        // When it has a selection set we are resolving an entity with a
-        // subselection. This can either be a list or an object.
-        const fieldSelect = getSelectionSet(node);
-
-        data[fieldAlias] = resolveResolverResult(
-          ctx,
-          resolverValue,
-          fieldKey,
-          fieldSelect,
-          data[fieldAlias] as Data | Data[]
-        );
-      }
-    } else if (node.selectionSet === undefined) {
+    if (node.selectionSet === undefined) {
       // The field is a scalar and can be retrieved directly
       if (fieldValue === undefined) {
         // Cache Incomplete: A missing field means it wasn't cached
@@ -169,55 +138,6 @@ const readSelection = (
   return data;
 };
 
-const resolveResolverResult = (
-  ctx: Context,
-  result: DataField,
-  key: string,
-  select: SelectionSet,
-  prevData: void | Data | Data[]
-) => {
-  // When we are dealing with a list we have to call this method again.
-  if (Array.isArray(result)) {
-    // @ts-ignore: Link cannot be expressed as a recursive type
-    return result.map((childResult, index) => {
-      const data = prevData !== undefined ? prevData[index] : undefined;
-      const indexKey = joinKeys(key, `${index}`);
-      return resolveResolverResult(ctx, childResult, indexKey, select, data);
-    });
-  } else if (result === null) {
-    return null;
-  } else if (isDataOrKey(result)) {
-    // We don't need to read the entity after exiting a resolver
-    // we can just go on and read the selection further.
-    const data = prevData === undefined ? Object.create(null) : prevData;
-    const childKey =
-      (typeof result === 'string' ? result : ctx.store.keyOfEntity(result)) ||
-      key;
-    const selectionResult = readSelection(ctx, childKey, select, data);
-
-    if (selectionResult !== null && typeof result === 'object') {
-      for (key in result) {
-        if (key !== '__typename' && result.hasOwnProperty(key)) {
-          selectionResult[key] = result[key];
-        }
-      }
-    }
-
-    return selectionResult;
-  }
-
-  warning(
-    false,
-    'Invalid resolver value: The resolver at `%s` returned a scalar (number, boolean, etc)' +
-      ', but the GraphQL query expects a selection set for this field.\n' +
-      'If necessary, use Cache.resolve() to resolve a link or entity from the cache.',
-    key
-  );
-
-  ctx.result.completeness = 'EMPTY';
-  return null;
-};
-
 const resolveLink = (
   ctx: Context,
   link: Link | Link[],
@@ -239,9 +159,3 @@ const resolveLink = (
     return readSelection(ctx, link, select, data);
   }
 };
-
-const isDataOrKey = (x: any): x is string | Data =>
-  typeof x === 'string' ||
-  (typeof x === 'object' &&
-    x !== null &&
-    typeof (x as any).__typename === 'string');
